Add -i ignore option for ignoring entries with specific mime types
This new option is very useful if wl-paste is used in combination with a clipboard manager. Often it is not desired to pass on sensitive data like a password to the clipboard-manager. Some password managers add specific MIME types to the password upon copying into the clipboard. This is an easy way to filter a specified type of data from being passed on.